move get_camera to camera base
