+++ multicamera_acquisition/interfaces/camera_base.py
@@ -1,9 +1,119 @@
-import numpy as np
 import yaml
+
 
 class CameraError(Exception):
     pass
 
+
+def get_camera(
+    brand="flir",
+    serial_number=None,
+    index=0,
+    config_file=None,
+):
+    """Get a camera object.
+    Parameters
+    ----------
+    brand : string (default: 'flir')
+        The brand of camera to use.  Currently only 'flir' is supported. If
+        'flir', the software PySpin is used. if 'basler', the software pypylon
+        is used.
+
+    serial_number : int or str (default: None)
+        The serial number of the camera. Ultimately used to find the index
+        of the camera in the software layer.
+        TAKES PRECEDENCE OVER INDEX.
+
+    index : int (default: 0)
+        The index of the camera to acquire, in a list of devices
+        enumerated by the software layer of the camera API.
+
+    config_file : string (default: None)
+        Path to a config file.  If None, the default config file for the given
+        camera brand will be used.
+
+    Returns
+    -------
+    cam : Camera object
+        The camera object, specific to the brand.
+
+    """
+    if brand == "flir":
+        from multicamera_acquisition.interfaces.camera_flir import FlirCamera as Camera
+
+        cam = Camera(index=str(index))
+
+        cam.init()
+
+        # # set gain
+        # cam.GainAuto = "Off"
+        # cam.Gain = gain
+
+        # # set exposure
+        # cam.ExposureAuto = "Off"
+        # cam.ExposureTime = exposure_time
+
+        # # set trigger
+        # if trigger == "arduino":
+        #     # TODO - many of these settings are not related to the trigger and should
+        #     # be redistributed
+        #     # TODO - remove hardcoding
+        #     cam.AcquisitionMode = "Continuous"
+        #     cam.AcquisitionFrameRateEnable = True
+        #     max_fps = cam.get_info("AcquisitionFrameRate")["max"]
+        #     cam.AcquisitionFrameRate = max_fps
+        #     cam.TriggerMode = "Off"
+        #     cam.TriggerSource = trigger_line
+        #     cam.TriggerOverlap = "ReadOut"
+        #     cam.TriggerSelector = "FrameStart"
+        #     cam.TriggerActivation = "RisingEdge"
+        #     # cam.TriggerActivation = "FallingEdge"
+        #     cam.TriggerMode = "On"
+
+        # else:
+        #     cam.LineSelector = trigger_line
+        #     cam.AcquisitionMode = "Continuous"
+        #     cam.TriggerMode = "Off"
+        #     cam.TriggerSource = "Software"
+        #     cam.V3_3Enable = True
+        #     cam.TriggerOverlap = "ReadOut"
+
+        # if roi is not None:
+        #     raise NotImplementedError("ROI not implemented for FLIR cameras")
+
+    elif brand == "basler":
+        from multicamera_acquisition.interfaces.camera_basler import BaslerCamera 
+        cam = BaslerCamera(index=index, config_file=config_file)
+        cam.init()
+
+    elif brand == "basler_emulated":
+        from multicamera_acquisition.interfaces.camera_basler import EmulatedBaslerCamera
+        cam = EmulatedBaslerCamera()
+        cam.init()
+
+    elif brand == "azure":
+        from multicamera_acquisition.interfaces.camera_azure import AzureCamera
+
+        if "name" in kwargs:
+            name = kwargs["name"]
+        else:
+            raise ValueError("Azure camera requires name")
+
+        cam = AzureCamera(
+            serial_number=str(serial), name=name, azure_index=kwargs["azure_index"]
+        )
+
+
+    elif brand == 'lucid':
+        from multicamera_acquisition.interfaces.camera_lucid import (
+            LucidCamera as Camera,
+        )
+        cam = Camera(
+            index=str(serial)
+        )
+        cam.init()
+
+    return cam
 
 class BaseCamera(object):
     """
